Separated endpoint constants and URL generators from the formatter. #7
