+++ src/main/groovy/com/xinra/growthlectures/frontend/Formatter.java
@@ -1,7 +1,5 @@
 package com.xinra.growthlectures.frontend;
 
-import com.xinra.growthlectures.service.LectureSummaryDto;
-import com.xinra.growthlectures.service.NamedDto;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import org.springframework.context.annotation.Scope;
@@ -10,10 +8,9 @@ import org.springframework.stereotype.Component;
 @Component
 @Scope("singleton")
 public class Formatter {
-
-  public static final String URL_CATEGORIES = "/lectures";
-  public static final String URL_LECTURERS = "/lecturers";
   
+  private final DateTimeFormatter dateFormatter =  DateTimeFormatter.ofPattern("dd.MM.yyyy");
+
   public String duration(Integer duration) {
     int hour = duration / 60 / 60;
     int minute = (duration / 60) % 60;
@@ -23,19 +20,7 @@ public class Formatter {
   }
   
   public String date(LocalDate date) {
-    return date.format(DateTimeFormatter.ofPattern("dd.MM.yyyy"));
-  }
-  
-  public String lectureUrl(LectureSummaryDto dto) {
-    return String.format("%s/%s/%s", URL_CATEGORIES, dto.getCategory().getSlug(), dto.getSlug());
-  }
-  
-  public String categoryUrl(NamedDto dto) {
-    return URL_CATEGORIES + "/" + dto.getSlug();
-  }
-  
-  public String lecturerUrl(NamedDto dto) {
-    return URL_LECTURERS + "/" + dto.getSlug();
+    return date.format(dateFormatter);
   }
   
 }
