+++ src/main/groovy/com/xinra/growthlectures/frontend/IndexController.java
@@ -22,7 +22,7 @@ public class IndexController {
   @Autowired
   private DtoFactory dtoFactory;
   
-  @RequestMapping("/")
+  @RequestMapping(Ui.URL_INDEX)
   public String index(Model model) {
     
     // PopularLecures
