+++ src/main/resources/templates/index.html
@@ -31,24 +31,24 @@
         <div class="panel panel-default">
           <div class="panel-heading">Popular Categories</div>
           <ul class="list-group">
-            <a th:each="popularCategory : ${popularCategories}" th:href="@{${@formatter.categoryUrl(popularCategory)}}" class="list-group-item">
+            <a th:each="popularCategory : ${popularCategories}" th:href="@{${@ui.categoryUrl(popularCategory)}}" class="list-group-item">
               <span th:text="${popularCategory.name}"></span> 
               <span th:text="${popularCategory.amount}" class="badge"></span>
             </a>
           </ul>
-          <div class="panel-footer"><a th:href="@{${@formatter.URL_CATEGORIES}}">Show all <span class="glyphicon glyphicon-chevron-right"></span></a></div>
+          <div class="panel-footer"><a th:href="@{${@ui.URL_CATEGORIES}}">Show all <span class="glyphicon glyphicon-chevron-right"></span></a></div>
         </div>
       </div>
       <div class="col-sm-6">
         <div class="panel panel-default">
           <div class="panel-heading">Popular Lecturers</div>
           <ul class="list-group">
-            <a th:each="popularLecturer : ${popularLecturers}" th:href="@{${@formatter.lecturerUrl(popularLecturer)}}" class="list-group-item">
+            <a th:each="popularLecturer : ${popularLecturers}" th:href="@{${@ui.lecturerUrl(popularLecturer)}}" class="list-group-item">
               <span th:text="${popularLecturer.name}"></span> 
               <span th:text="${popularLecturer.amount}" class="badge"></span>
             </a>
           </ul>
-          <div class="panel-footer"><a th:href="@{${@formatter.URL_LECTURERS}}">Show all <span class="glyphicon glyphicon-chevron-right"></span></a></div>
+          <div class="panel-footer"><a th:href="@{${@ui.URL_LECTURERS}}">Show all <span class="glyphicon glyphicon-chevron-right"></span></a></div>
         </div>
       </div>
     </div>
